Move BrowseArgs into search models
