+++ app/api/api_v1/routers/summaries.py
@@ -10,9 +10,9 @@ from db_client.models.dfce import FamilyCategory, Geography
 from fastapi import APIRouter, Depends, HTTPException, Request, status
 
 from app.clients.db.session import get_db
-from app.models.search import GeographySummaryFamilyResponse
+from app.models.search import BrowseArgs, GeographySummaryFamilyResponse
 from app.repository.lookups import get_country_slug_from_country_code, is_country_code
-from app.repository.search import BrowseArgs, browse_rds_families
+from app.repository.search import browse_rds_families
 
 _LOGGER = logging.getLogger(__name__)
