+++ app/models/search.py
@@ -365,3 +365,17 @@ class GeographySummaryFamilyResponse(BaseModel):
     family_counts: Mapping[FamilyCategory, int]
     top_families: Mapping[FamilyCategory, _T5FamL]
     targets: Sequence[str]  # TODO: Placeholder for later
+
+
+class BrowseArgs(BaseModel):
+    """Arguments for the browse_rds function"""
+
+    geography_slugs: Optional[Sequence[str]] = None
+    country_codes: Optional[Sequence[str]] = None
+    start_year: Optional[int] = None
+    end_year: Optional[int] = None
+    categories: Optional[Sequence[str]] = None
+    sort_field: SortField = SortField.DATE
+    sort_order: SortOrder = SortOrder.DESCENDING
+    offset: Optional[int] = 0
+    limit: Optional[int] = 10
